+++ app/src/main/java/com/example/meetup/model/CategoryIdResponseModel.kt
@@ -12,8 +12,8 @@ data class Food(
     val name: String,
     val storeId: Long,
     val storeName: String,
-    val dollarPrice: Long,
-    val canadaPrice: Long,
+    val dollarPrice: Double,
+    val canadaPrice: Double,
     val image: String?,
-    val avgRate: Long,
+    val avgRate: Double,
 )
