+++ app/src/main/java/com/example/meetup/model/MenuAddRequestModel.kt
@@ -9,8 +9,8 @@ val informationImage: String,
 data class MenuAddRequestModelDto(
     val categoryId: Long,
     val name: String,
-    val dollarPrice: Long,
-    val canadaPrice: Long,
+    val dollarPrice: Double,
+    val canadaPrice: Double,
     val description: String,
     val foodOptionRequestList: ArrayList<MenuAddRequestModelDtoFoodOptionRequestList>,
     val foodPackage: String,
@@ -19,7 +19,7 @@ data class MenuAddRequestModelDto(
 
 data class MenuAddRequestModelDtoFoodOptionRequestList(
     val name: String,
-    val dollarPrice: Long,
-    val canadaPrice: Long,
+    val dollarPrice: Double,
+    val canadaPrice: Double,
 )
